# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Nula Logistics API documentation",
        "description": "API documentation for Products, Suppliers and Warehouses management. All routes start with /api namespace",
        "contact": {
            "email": "katerinka.melnichuk@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/suppliers": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Retrieve all suppliers",
                "operationId": "a043e760500e8b8058d41738e74046f7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string"
                                            },
                                            "updated_at": {
                                                "type": "string"
                                            },
                                            "products": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "supplier_id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "price": {
                                                            "type": "number"
                                                        },
                                                        "created_at": {
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "warehouses": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "supplier_id": {
                                                            "type": "integer"
                                                        },
                                                        "address": {
                                                            "type": "string"
                                                        },
                                                        "product_count": {
                                                            "type": "number"
                                                        },
                                                        "created_at": {
                                                            "type": "string"
                                                        },
                                                        "updated_at": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Suppliers"
                ],
                "summary": "Create new supplier",
                "description": "Create a new supplier in the database",
                "operationId": "3e424442ff493b5f7620e1ae344a23dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Supplier's name",
                                        "type": "string",
                                        "example": "Supplier A"
                                    },
                                    "address": {
                                        "description": "Supplier's address",
                                        "type": "string",
                                        "example": "123 Main St"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Supplier's ID",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Supplier's name",
                                            "type": "string"
                                        },
                                        "address": {
                                            "description": "Supplier's address",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    }
}